requirements.txt
python-telegram-bot==13.15
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.0.1 